chore: refactor and cleanup
