+++ packages/eslint-plugin-lit-a11y/lib/rules/aria-attr-valid-value.js
@@ -6,7 +6,6 @@
 const { aria } = require('aria-query');
 const { TemplateAnalyzer } = require('../../template-analyzer/template-analyzer.js');
 const { getElementAriaAttributes } = require('../utils/aria.js');
-const { getAttrVal } = require('../utils/getAttrVal.js');
 const { isHtmlTaggedTemplate } = require('../utils/isLitHtmlTemplate.js');
 const { hasLitHtmlImport, createValidLitHtmlSources } = require('../utils/utils.js');
 
@@ -129,7 +128,7 @@ const AriaAttrTypesRule = {
 
               if (ariaAttributes.length > 0) {
                 ariaAttributes.forEach(attr => {
-                  const val = getAttrVal(element.attribs[attr]);
+                  const val = element.attribs[attr];
 
                   // Ignore the attribute if its value is null or undefined.
                   if (val == null) return;
