+++ packages/eslint-plugin-lit-a11y/lib/utils/getExpressionValue.js
@@ -1,0 +1,13 @@
+function getExpressionValue(analyzer, placeholder) {
+  let val = analyzer.expressionValues.get(placeholder);
+
+  if (typeof val === 'string') {
+    val = val.replace(/"|'/g, '');
+  }
+
+  return val;
+}
+
+module.exports = {
+  getExpressionValue,
+};
