+++ packages/eslint-plugin-lit-a11y/tests/lib/rules/img-redundant-alt.js
@@ -67,7 +67,7 @@ ruleTester.run('img-redundant-alt', rule, {
       code: 'html`<img src="baz" alt=${"photo of dog"} />`',
       errors: [
         {
-          messageId:'imgRedundantAlt',
+          messageId: 'imgRedundantAlt',
           data: { banned: 'photo', plural: 'word' },
         },
       ],
